+++ src/vec2.rs
@@ -4,12 +4,12 @@ use wasm_bindgen::prelude::wasm_bindgen;
 // Use copy instead of clone
 #[wasm_bindgen]
 pub struct Vec2 {
-    pub x: usize,
-    pub y: usize,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Vec2 {
-    pub fn new(x: usize, y: usize) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Vec2 { x, y }
     }
 }
